training/requirements_train.txt
torch>=2.4.0
torchvision>=0.19.0
accelerate>=0.34.0
diffusers>=0.31.0
transformers>=4.44.0
datasets>=2.20.0
huggingface_hub>=0.24.0
open_clip_torch>=3.2.0
insightface>=0.7.3
onnxruntime-gpu>=1.16.0
safetensors>=0.4.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
